Revert "Allow array-refs to be quoted as IN-style lists"
This reverts commit 361e516e346042f22060b01370242713eac30bf6.

I've come to the conclusion that this was a mistake: it's incompatible with
server-side prepared queries, it goes against DBI rules, and it's likely to
impede any future attempt to handle PostgreSQL arrays in a way that
integrates well with Perl.  Fortunately, I was sensible enough even five
years ago to avoid documenting this change.  So if you used it, sorry, but
that's what you get for relying on undocumented features.