Delete Japanese-language documentation
I can't read it, so I have no idea whether it's correct, and it wasn't
readable on search.cpan.org anyway.

In principle, I'd have nothing against including documentation in languages
other than English, as long as a fluent speaker of the language in question
were volunteering to maintain the docs.