Require Perl 5.8
I no longer seem to be able to build a working Perl 5.6.2 on my main
development system, and since Perl 5.12 is just around the corner, it
strikes me as a bit silly to go to great lengths to support 5.6.

This also means there are no dependency-management issues for requiring
Digest::MD5, because that module ships with Perl 5.8 and newer.